refactor: all-stories page with components
references #344
